Component manager tests passing
